rbot-office/requirements.txt
slack_bolt
python-dotenv
google-genai
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
ddgs